wdio-cli: move out code to constants and utils
