+++ packages/wdio-cli/src/constants.js
@@ -1,0 +1,72 @@
+const DEFAULT_TIMEOUT = 10000
+const NOOP = function () {}
+
+export const DEFAULT_CONFIGS = {
+    sync: true,
+    specs: [],
+    suites: {},
+    exclude: [],
+    logLevel: 'silent',
+    coloredLogs: true,
+    deprecationWarnings: true,
+    baseUrl: null,
+    bail: 0,
+    waitforInterval: 500,
+    waitforTimeout: 1000,
+    framework: 'mocha',
+    reporters: [],
+    reporterOptions: {},
+    maxInstances: 100,
+    maxInstancesPerCapability: 100,
+    connectionRetryTimeout: 90000,
+    connectionRetryCount: 3,
+    debug: false,
+    execArgv: null,
+
+    /**
+     * framework defaults
+     */
+    mochaOpts: {
+        timeout: DEFAULT_TIMEOUT
+    },
+    jasmineNodeOpts: {
+        defaultTimeoutInterval: DEFAULT_TIMEOUT
+    },
+
+    /**
+     * hooks
+     */
+    onPrepare: NOOP,
+    before: [],
+    beforeSession: [],
+    beforeSuite: [],
+    beforeHook: [],
+    beforeTest: [],
+    beforeCommand: [],
+    afterCommand: [],
+    afterTest: [],
+    afterHook: [],
+    afterSuite: [],
+    afterSession: [],
+    after: [],
+    onComplete: NOOP,
+    onError: [],
+    onReload: [],
+
+    /**
+     * cucumber specific hooks
+     */
+    beforeFeature: [],
+    beforeScenario: [],
+    beforeStep: [],
+    afterFeature: [],
+    afterScenario: [],
+    afterStep: []
+}
+
+export const SUPPORTED_HOOKS = [
+    'before', 'beforeSession', 'beforeSuite', 'beforeHook', 'beforeTest', 'beforeCommand',
+    'afterCommand', 'afterTest', 'afterHook', 'afterSuite', 'afterSession', 'after',
+    'beforeFeature', 'beforeScenario', 'beforeStep', 'afterFeature',
+    'afterScenario', 'afterStep', 'onError', 'onReload'
+]
